+++ diff_converter.h
@@ -17,7 +17,7 @@ public:
     DiffConverter() noexcept;
     explicit DiffConverter(std::unordered_map<unsigned int, std::string> &&difference_map_arg);
     boost::optional<std::string> get_char(char c) const;
-    std::pair<std::string, double> get_string(const std::string &s, const Fonts &fonts) const;
+    std::pair<std::string, float> get_string(const std::string &s, const Fonts &fonts) const;
     bool is_empty() const;
     static DiffConverter get_converter(const dict_t &dictionary,
                                        const std::pair<std::string, pdf_object_t> &differences,
